3 Different Functions to Detect Duplicate value
3 Different Functions to detect the Duplicate value in any Array.

First function works with 2 pointers( i , j ), checks if the value of both indexes is the same or not.

Second function works with 2 different arrays other than one that will be given to the function as an argument. Take the value one after another from the original array and push it in one array only when the same value isn't available in that array. And if the value is already present there, then it should maintain the duplicate array and pass the value in it too. The fun part is, the function will also make sure that there isn't duplicate value inside the duplicate array. ( consider the case where you have the same number more than 2 times)

Third Function's functionality is based on the second function but the difference is this time function is not iterative but Recursive, and the rest of the functionality is the same.